requirements.txt
Django>=5.0,<6.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-spectacular
psycopg2-binary
dj-database-url
whitenoise
gunicorn
